+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rich text for `alert message`, `bullet list` components (Orange-OpenSource/ouds-ios#1405)
 - Rich text for `alert message`, `bullet list`, `switch item` components (Orange-OpenSource/ouds-ios#1405)
 - Rich text for `alert message`, `bullet list`, `switch item`, `radio item` components (Orange-OpenSource/ouds-ios#1405)
+- Rich text support for `alert message`, `bullet list`, `switch item`, `radio item`, `checkbox item` components (Orange-OpenSource/ouds-ios#1405)
 
 ### Fixed